+++ mmd/pmx_writer.go
@@ -29,7 +29,7 @@ func (p *baseWriter) writeInt(v int) {
 	binary.Write(p.w, binary.LittleEndian, &vv)
 }
 
-func (p *baseWriter) writeFloat(v float32) float32 {
+func (p *baseWriter) writeFloat(v float64) float64 {
 	binary.Write(p.w, binary.LittleEndian, &v)
 	return v
 }
